+++ src/components/Auth/PasswordReset/index.jsx
@@ -1,8 +1,7 @@
-import React, { Component } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import styles from './style.module.css';
-import { withFirebase } from '../../../services/firebase';
 import { Link } from 'react-router-dom';
-
 import * as ROUTES from '../../../constants/routes';
 
 import Container from '@material-ui/core/Container';
@@ -13,177 +12,103 @@ import InputField from '../../UI/InputField';
 import CustomButton from '../../UI/CustomButton';
 import Progress from '../../UI/Progress';
 
-import {
-  validateEmail,
-  validatePasswordResetForm
-} from '../../../utils/validate';
+const PasswordReset = props => {
+  const {
+    validateEmail,
+    sendEmail,
+    emailError,
+    isProcessing,
+    successMessage,
+    errorMessage,
+    dismissMessage
+  } = props;
 
-class PasswordReset extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      email: null,
-      emailError: null,
-      isProcessing: false,
-      successMessage: null,
-      errorMessage: null
-    };
-
-    this.validateEmail = this.validateEmail.bind(this);
-    this.sendEmail = this.sendEmail.bind(this);
-    this.dismissMessage = this.dismissMessage.bind(this);
-  }
-
-  validateEmail(value) {
-    const result = validateEmail(value);
-    // console.log(result);
-
-    if (result.isValid !== true) {
-      this.setState({ emailError: result.message });
-    } else {
-      this.setState({ emailError: null, email: value });
-    }
-  }
-
-  sendEmail() {
-    this.setState({
-      isProcessing: true
-    });
-
-    const result = validatePasswordResetForm();
-
-    let email = true;
-    let emailError = null;
-
-    if (result.isValid !== true) {
-      email = false;
-      emailError = result.message;
-    }
-
-    if (email === false) {
-      this.setState({
-        emailError,
-        isProcessing: false
-      });
-      return false;
-    }
-
-    setTimeout(() => {
-      const { firebase } = this.props;
-
-      firebase
-        .passwordReset(this.state.email)
-        .then(() => {
-          this.setState({
-            successMessage:
-              'Password reset link has been send to your provided email address, check you mailbox.',
-            errorMessage: null,
-            isProcessing: false
-          });
-        })
-        .catch(error => {
-          const errorMessage = error.message;
-          console.log(errorMessage);
-
-          this.setState({
-            successMessage: null,
-            errorMessage,
-            isProcessing: false
-          });
-        });
-    }, 3000);
-  }
-
-  dismissMessage() {
-    this.setState({ successMessage: null, errorMessage: null });
-  }
-
-  render() {
-    const {
-      emailError,
-      isProcessing,
-      successMessage,
-      errorMessage
-    } = this.state;
-
-    return (
-      <Container className={styles.container}>
-        {isProcessing && <Progress />}
-        <Paper className="root">
-          <Grid container spacing={1}>
-            <Grid item xs={12}>
-              <Typography variant="h1" className={styles.title}>
-                Forgot your Password?
+  return (
+    <Container className={styles.container}>
+      {isProcessing && <Progress />}
+      <Paper className="root">
+        <Grid container spacing={1}>
+          <Grid item xs={12}>
+            <Typography variant="h1" className={styles.title}>
+              Forgot your Password?
+            </Typography>
+          </Grid>
+          {(successMessage || errorMessage) && (
+            <Grid
+              item
+              xs={12}
+              className={successMessage ? styles.success : styles.error}
+            >
+              <Typography variant="overline">
+                {successMessage ? successMessage : errorMessage}
+                <CustomButton
+                  // type="secondary"
+                  disableRipple={true}
+                  onClick={dismissMessage}
+                >
+                  Dismiss
+                </CustomButton>
               </Typography>
             </Grid>
-            {(successMessage || errorMessage) && (
-              <Grid
-                item
-                xs={12}
-                className={successMessage ? styles.success : styles.error}
-              >
-                <Typography variant="overline">
-                  {successMessage ? successMessage : errorMessage}
+          )}
+          <Grid item xs={12}>
+            <InputField
+              focus={true}
+              label="Email"
+              type="email"
+              validate={validateEmail}
+              errorMessage={emailError}
+              disabled={isProcessing}
+            />
+          </Grid>
+          <Grid item xs={12}>
+            <Typography variant="subtitle1" align="center">
+              We'll send you a link to reset your password.
+            </Typography>
+          </Grid>
+          <Grid container className={styles.btnWrapper}>
+            <Grid item xs={6}>
+              <div className={styles.btnWrapperChild1}>
+                <Link to={ROUTES.SIGNIN} className={styles.link}>
                   <CustomButton
+                    variant="outlined"
                     // type="secondary"
-                    disableRipple={true}
-                    onClick={this.dismissMessage}
-                  >
-                    Dismiss
-                  </CustomButton>
-                </Typography>
-              </Grid>
-            )}
-            <Grid item xs={12}>
-              <InputField
-                focus={true}
-                label="Email"
-                type="email"
-                validate={this.validateEmail}
-                errorMessage={emailError}
-                disabled={isProcessing}
-              />
-            </Grid>
-            <Grid item xs={12}>
-              <Typography variant="subtitle1" align="center">
-                We'll send you a link to reset your password.
-              </Typography>
-            </Grid>
-            <Grid container className={styles.btnWrapper}>
-              <Grid item xs={6}>
-                <div className={styles.btnWrapperChild1}>
-                  <Link to={ROUTES.SIGNIN} className={styles.link}>
-                    <CustomButton
-                      variant="outlined"
-                      // type="secondary"
-                      size="large"
-                      disabled={isProcessing}
-                    >
-                      Back to Sign In
-                    </CustomButton>
-                  </Link>
-                </div>
-              </Grid>
-              <Grid item xs={6}>
-                <div className={styles.btnWrapperChild2}>
-                  <CustomButton
-                    variant="contained"
-                    type="primary"
                     size="large"
-                    onClick={this.sendEmail}
                     disabled={isProcessing}
                   >
-                    Email me Reset Link
+                    Back to Sign In
                   </CustomButton>
-                </div>
-              </Grid>
+                </Link>
+              </div>
+            </Grid>
+            <Grid item xs={6}>
+              <div className={styles.btnWrapperChild2}>
+                <CustomButton
+                  variant="contained"
+                  type="primary"
+                  size="large"
+                  onClick={sendEmail}
+                  disabled={isProcessing}
+                >
+                  Email me Reset Link
+                </CustomButton>
+              </div>
             </Grid>
           </Grid>
-        </Paper>
-      </Container>
-    );
-  }
-}
+        </Grid>
+      </Paper>
+    </Container>
+  );
+};
 
-export default withFirebase(PasswordReset);
+PasswordReset.propTypes = {
+  validateEmail: PropTypes.func.isRequired,
+  sendEmail: PropTypes.func.isRequired,
+  emailError: PropTypes.string.isRequired,
+  isProcessing: PropTypes.bool.isRequired,
+  successMessage: PropTypes.string.isRequired,
+  errorMessage: PropTypes.string.isRequired,
+  dismissMessage: PropTypes.func.isRequired
+};
+
+export default PasswordReset;
